backend/requirements.txt
python-dotenv
sentence-transformers
langchain
langchain-community
langchain-text-splitters
langchain-chroma
pydantic
pydantic-settings
groq
PyPDF2
